requirements.txt
# CPU Scheduler Simulation Dependencies

# Core dependencies
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
scipy>=1.7.0

# Configuration
pyyaml>=6.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Documentation
jupyter>=1.0.0
ipython>=8.0.0

# Development tools
black>=22.0.0
flake8>=5.0.0
mypy>=0.991